ResourceOperationRequired added

+++ RestaurantAPI_5.0/Services/RestaurantService.cs
@@ -1,11 +1,15 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using RestaurantAPI_5._0.Authorization;
 using RestaurantAPI_5._0.Entities;
 using RestaurantAPI_5._0.Exceptions;
 using RestaurantAPI_5._0.Models;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 namespace RestaurantAPI_5._0.Services
 {
@@ -15,19 +19,31 @@ namespace RestaurantAPI_5._0.Services
     {
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper)
+        private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper,ILogger<RestaurantService> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
+            _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Update(int id, UpdateRestaurantDto dto)
+        public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
         {
+            
             var restaurant = _dbContext
                .Restaurants
                .FirstOrDefault(r => r.Id == id);
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException(); 
+            }
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
             restaurant.HasDelivery = dto.HasDelivery;
@@ -35,15 +51,21 @@ namespace RestaurantAPI_5._0.Services
             _dbContext.SaveChanges();
         }
 
-        public void Delete(int id)
+        public void Delete(int id,ClaimsPrincipal user)
         {
+            _logger.LogError($"Restaurant with id: {id} DELETE action invoked");
             var restaurant = _dbContext
                   .Restaurants
                   .FirstOrDefault(r => r.Id == id);
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
-
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+               new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
         }
@@ -70,9 +92,11 @@ namespace RestaurantAPI_5._0.Services
             var restaurantsDtos = _mapper.Map<List<RestaurantDto>>(restaurants);
             return restaurantsDtos;
         }
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto,int userId)
         {
+            
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
             return restaurant.Id;
